wip - refactor user controller, service, and add validations

+++ MrTakuVetClinic/Services/UserService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using MrTakuVetClinic.Entities;
 using MrTakuVetClinic.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -9,10 +10,12 @@ namespace MrTakuVetClinic.Services
     public class UserService
     {
         private readonly IUserRepository _userRepository;
+        private readonly IUserTypeRepository _userTypeRepository;
 
-        public UserService(IUserRepository userRepository)
+        public UserService(IUserRepository userRepository, IUserTypeRepository userTypeRepository)
         {
             _userRepository = userRepository;
+            _userTypeRepository = userTypeRepository;
         }
 
         public async Task<IEnumerable<User>> GetAllUsersAsync()
@@ -22,7 +25,12 @@ namespace MrTakuVetClinic.Services
 
         public async Task<User> GetUserByUsernameAsync(string username)
         {
-            return await _userRepository.GetUserByUsernameAsync(username);
+            var user = await _userRepository.GetUserByUsernameAsync(username);
+            if (user == null)
+            {
+                throw new ArgumentException("User not found.");
+            }
+            return user;
         }
 
         public async Task<IEnumerable<User>> GetSearchUsersAsync([FromQuery] string firstName, string lastName)
@@ -32,6 +40,11 @@ namespace MrTakuVetClinic.Services
 
         public async Task AddUserAsync(User user)
         {
+            var userType = await _userTypeRepository.GetByIdAsync(user.UserTypeId);
+            if (userType == null)
+            {
+                throw new ArgumentException("User type does not exist.");
+            }
             await _userRepository.AddAsync(user);
         }
 
@@ -42,6 +55,10 @@ namespace MrTakuVetClinic.Services
 
         public async Task DeleteUserByUsernameAsync(string username)
         {
+            if (await _userRepository.GetUserByUsernameAsync(username) == null)
+            {
+                throw new ArgumentException("User not found.");
+            }
             await _userRepository.DeleteUserByUsernameAsync(username);
         }
     }
